refactor(compiler): move AnyToExistsTransform to expression construction time
