+++ ibis/expr/types/logical.py
@@ -75,7 +75,10 @@ class BooleanValue(NumericValue):
     def __invert__(self) -> BooleanValue:
         from ibis.expr import operations as ops
 
-        return ops.Not(self).to_expr()
+        try:
+            return self.negate()
+        except AttributeError:
+            return ops.Not(self).to_expr()
 
 
 @public
